[FIR] Make getOrDefault in JvmMappedScope low-priority instead of forbidding it
Generally we now have two 'getOrDefault' in Kotlin classes,
the first is taken from Java Map and
the second one is taken from Kotlin Map.
The forbidding solution may (? no test case ?) break something.
However, the low priority solution in this commit is also questionable,
because we have also not to match two these functions as overrides,
and then not to generate low-priority fake override to avoid clashing.